backend/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
azure-ai-openai